+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 0.4.0 [2023-10-05]
 
 - Adds the settings page.
 - Adds a fake login button.
@@ -20,7 +20,7 @@
 - Adds Isar.
 - Adds theme toggle buttons.
 
-## 0.1.0 [2023-09-28]
+## 0.1.0 [2023-09-28
 
 - Adds `go_router`'s router.
 - Adds dummy events and users screens.